[Test] Fix D2H window CPU classification
